Split meta commands into modules
