+++ src/busybox/meta/downgrade.rs
@@ -1,0 +1,6 @@
+use ::{Result, Shell, Arguments};
+
+pub fn downgrade(sh: &mut Shell, _args: Arguments) -> Result<()> {
+    sh.downgrade();
+    Ok(())
+}
